Move remaining utilities
