+++ src/main/scala/scripts/PhysSimReplayer.scala
@@ -1,10 +1,7 @@
-package beam.physsim.jdeqsim
-
-import java.io.{Closeable, File}
-import java.nio.charset.StandardCharsets
-import java.nio.file.Files
+package scripts
 
 import beam.agentsim.events.PathTraversalEvent
+import beam.physsim.jdeqsim.AgentSimToPhysSimPlanConverter
 import beam.sim.{BeamConfigChangesObservable, BeamHelper}
 import beam.utils.{BeamConfigUtils, EventReader}
 import com.typesafe.config.{Config, ConfigFactory, ConfigResolveOptions, ConfigValueFactory}
@@ -13,6 +10,9 @@ import org.matsim.api.core.v01.events.Event
 import org.matsim.core.controler.events.IterationEndsEvent
 import org.matsim.core.events.EventsManagerImpl
 
+import java.io.{Closeable, File}
+import java.nio.charset.StandardCharsets
+import java.nio.file.Files
 import scala.io.Source
 import scala.util.Try
