+++ src/main/scala/scripts/R5RoutingApp.scala
@@ -1,6 +1,5 @@
-package beam.router
+package scripts
 
-import java.util.concurrent.TimeUnit
 import akka.actor.{ActorRef, ActorSystem, Identify, Props}
 import akka.http.scaladsl.Http
 import akka.http.scaladsl.model.{HttpResponse, StatusCodes}
@@ -13,6 +12,7 @@ import beam.agentsim.agents.vehicles.VehicleProtocol.StreetVehicle
 import beam.agentsim.events.SpaceTime
 import beam.router.BeamRouter.{Location, RoutingRequest, RoutingResponse, UpdateTravelTimeLocal}
 import beam.router.Modes.BeamMode.CAR
+import beam.router.RoutingWorker
 import beam.sim.config.BeamConfig
 import beam.sim.{BeamHelper, BeamWarmStart}
 import beam.utils.{DateUtils, FileUtils, LoggingUtil}
@@ -21,6 +21,7 @@ import de.heikoseeberger.akkahttpcirce.FailFastCirceSupport
 import io.circe.syntax._
 import org.matsim.api.core.v01.Id
 
+import java.util.concurrent.TimeUnit
 import scala.concurrent.Await
 import scala.concurrent.duration._
